+++ core/src/main/java/nl/weeaboo/vn/impl/core/Novel.java
@@ -84,7 +84,6 @@ public class Novel implements INovel {
         // Clear caches
         StaticEnvironment.TEXTURE_STORE.get().clear();
         StaticEnvironment.SHADER_STORE.get().clear();
-        StaticEnvironment.MUSIC_STORE.get().clear();
         StaticEnvironment.FONT_STORE.get().clear();
 
         start(KnownScriptFunctions.TITLESCREEN);