+++ core/src/main/java/nl/weeaboo/vn/impl/sound/INativeAudioFactory.java
@@ -3,6 +3,8 @@ package nl.weeaboo.vn.impl.sound;
 import java.io.IOException;
 import java.io.Serializable;
 
+import com.badlogic.gdx.audio.Music;
+
 import nl.weeaboo.filesystem.FilePath;
 
 /**
@@ -12,5 +14,7 @@ interface INativeAudioFactory extends Serializable {
 
     /** Creates a new {@link INativeAudio} instance. */
     INativeAudio createNativeAudio(FilePath filePath) throws IOException;
+
+    Music newGdxMusic(FilePath filePath);
 
 }