+++ core/src/main/java/nl/weeaboo/vn/impl/sound/NativeAudioFactory.java
@@ -1,12 +1,11 @@
 package nl.weeaboo.vn.impl.sound;
 
-import java.io.FileNotFoundException;
-import java.io.IOException;
-
+import com.badlogic.gdx.Gdx;
+import com.badlogic.gdx.assets.AssetManager;
+import com.badlogic.gdx.assets.loaders.FileHandleResolver;
 import com.badlogic.gdx.audio.Music;
 
 import nl.weeaboo.filesystem.FilePath;
-import nl.weeaboo.vn.gdx.res.IResource;
 import nl.weeaboo.vn.impl.core.StaticEnvironment;
 import nl.weeaboo.vn.impl.core.StaticRef;
 
@@ -14,15 +13,17 @@ final class NativeAudioFactory implements INativeAudioFactory {
 
     private static final long serialVersionUID = SoundImpl.serialVersionUID;
 
-    private final StaticRef<GdxMusicStore> musicStore = StaticEnvironment.MUSIC_STORE;
+    private final StaticRef<AssetManager> assetManager = StaticEnvironment.ASSET_MANAGER;
 
     @Override
-    public INativeAudio createNativeAudio(FilePath filePath) throws IOException {
-        IResource<Music> resource = musicStore.get().get(filePath);
-        if (resource == null) {
-            throw new FileNotFoundException("Unable to load resource: " + filePath);
-        }
-        return new NativeAudio(resource);
+    public INativeAudio createNativeAudio(FilePath filePath) {
+        return new NativeAudio(this, filePath);
+    }
+
+    @Override
+    public Music newGdxMusic(FilePath filePath) {
+        FileHandleResolver fileResolver = assetManager.get().getFileHandleResolver();
+        return Gdx.audio.newMusic(fileResolver.resolve(filePath.toString()));
     }
 
 }