+++ core/src/main/java/nl/weeaboo/vn/impl/sound/SoundImpl.java
@@ -2,7 +2,7 @@ package nl.weeaboo.vn.impl.sound;
 
 final class SoundImpl {
 
-    static final long serialVersionUID = 4L;
+    static final long serialVersionUID = 5L;
 
     private SoundImpl() {
     }