+++ core/src/main/lua/builtin/script/vn/sound.lua
@@ -14,6 +14,7 @@ local function tosound(soundOrChannel)
 end
 
 local MUSIC_CHANNEL = 9000
+local DEFAULT_SOUND_CHANNEL = 9100
 
 --Default fade-out time when stopping music, can be changed using setDefaultMusicFadeTime()
 local defaultMusicStopFrames = 60
@@ -44,7 +45,10 @@ function musicStop(fadeTimeFrames)
     if s == nil then
         return
     end
-    return soundStop(s, fadeTimeFrames or defaultMusicStopFrames)
+    if fadeTimeFrames == nil then
+        fadeTimeFrames = defaultMusicStopFrames
+    end
+    return soundStop(s, fadeTimeFrames)
 end
 
 ---Stops background music started with the <code>music</code> function, blocking until the music has finished stopping.
@@ -53,12 +57,14 @@ end
 -- @see music
 -- @see musicStop
 function musicStopAndWait(fadeTimeFrames)
-    fadeTimeFrames = fadeTimeFrames or defaultMusicStopFrames
+    if fadeTimeFrames == nil then
+        fadeTimeFrames = defaultMusicStopFrames
+    end
 
     local s = musicStop(fadeTimeFrames)
     if s ~= nil and fadeTimeFrames > 0 then
         -- Wait for music top finish stopping
-        yield(fadeTimeFrames)
+        yield(fadeTimeFrames / getEffectSpeed())
     end
 end
 
@@ -95,14 +101,18 @@ end
 -- @treturn ISound A sound object with which to control playback, or <code>nil</code> if for whatever reason
 --          no sound could be started.
 function sound(filename, loops, volume, channel, type)
-    loops = loops or 1
     if loops == 0 then
         --Playing a sound zero times is easy, just don't do anything
         return
     end
-    
-    volume = volume or 1.0
-    channel = channel or 9100
+    loops = loops or 1
+
+    if volume == nil then
+        volume = volume or 1.0
+    end
+    if channel == nil then
+        channel = DEFAULT_SOUND_CHANNEL
+    end
     type = type or SoundType.SOUND
 
     local s = Sound.create(filename, type)
@@ -115,20 +125,20 @@ function sound(filename, loops, volume, channel, type)
 end
 
 ---Stops a playing sound.
--- @param soundOrChannel The Sound object or audio channel to change the volume of.
+-- @param soundOrChannel[opt=9100] The Sound object or audio channel to stop.
 -- @number fadeTimeFrames Optional argument specifying the duration (in frames)
 --         of a slow fade-out instead of stopping playback immediately.
 -- @return The sound which had stopped, or is now in the process of stopping
 -- @see sound
 function soundStop(soundOrChannel, fadeTimeFrames)
-    local sound = tosound(soundOrChannel)    
+    local sound = tosound(soundOrChannel or DEFAULT_SOUND_CHANNEL)
     if sound == nil then
         Log.warn("Attempt to stop invalid sound effect or music: {}", soundOrChannel)
         return
     end
 
     fadeTimeFrames = fadeTimeFrames or 0
-    sound:stop(fadeTimeFrames)
+    sound:stop(fadeTimeFrames / getEffectSpeed())
     return sound
 end
 
@@ -144,9 +154,11 @@ function changeVolume(soundOrChannel, targetVolume, durationFrames)
         return
     end
 
-    targetVolume = targetVolume or 1
+    if targetVolume == nil then
+        targetVolume = 1
+    end
     durationFrames = durationFrames or 0
-    
+
     local vol = sound:getPrivateVolume()
     if durationFrames > 0 then
         local delta = (targetVolume - vol) / durationFrames