+++ core/src/test/java/nl/weeaboo/vn/impl/core/TestEnvironment.java
@@ -15,7 +15,6 @@ import nl.weeaboo.vn.impl.script.lib.BasicScriptInitializer;
 import nl.weeaboo.vn.impl.script.lua.LuaScriptEnv;
 import nl.weeaboo.vn.impl.script.lua.LuaScriptLoader;
 import nl.weeaboo.vn.impl.script.lua.LuaTestUtil;
-import nl.weeaboo.vn.impl.sound.GdxMusicStore;
 import nl.weeaboo.vn.impl.sound.SoundModule;
 import nl.weeaboo.vn.impl.stats.PlayTimerStub;
 import nl.weeaboo.vn.impl.stats.StatsModule;
@@ -59,7 +58,6 @@ public class TestEnvironment extends DefaultEnvironment {
                 gdxFileSystem, prefs));
         StaticEnvironment.SHADER_STORE.set(new ShaderStore());
         StaticEnvironment.FONT_STORE.set(new GdxFontStore(gdxFileSystem));
-        StaticEnvironment.MUSIC_STORE.set(new GdxMusicStore(StaticEnvironment.MUSIC_STORE));
 
         TestEnvironment env = new TestEnvironment(input);
         env.renderEnv = CoreTestUtil.BASIC_ENV;